git-svn-id: https://geonetwork.svn.sourceforge.net/svnroot/geonetwork/trunk@2752 ff403467-3f20-0410-99e5-ff581e626a08

+++ src/org/fao/geonet/services/resources/Upload.java
@@ -24,13 +24,16 @@
 package org.fao.geonet.services.resources;
 
 import java.io.File;
+import java.util.Iterator;
 import jeeves.interfaces.Service;
-import jeeves.server.ServiceConfig;
+import jeeves.resources.dbms.Dbms;
 import jeeves.server.context.ServiceContext;
+import jeeves.server.ServiceConfig;
+import jeeves.server.UserSession;
 import jeeves.utils.Util;
+import org.fao.geonet.constants.Geonet;
 import org.fao.geonet.constants.Params;
 import org.fao.geonet.lib.Lib;
-import org.fao.geonet.services.metadata.Update;
 import org.jdom.Element;
 
 //=============================================================================
@@ -41,35 +44,56 @@ import org.jdom.Element;
 public class Upload implements Service
 {
 	private Element config;
-	private Update  update = new Update();
 
-	//-----------------------------------------------------------------------------
+	//----------------------------------------------------------------------------
 	//---
 	//--- Init
 	//---
-	//-----------------------------------------------------------------------------
+	//----------------------------------------------------------------------------
 
-	public void init(String appPath, ServiceConfig params) throws Exception
-	{
-		update.init(appPath, params);
-	}
+	public void init(String appPath, ServiceConfig params) throws Exception {}
 
-	//-----------------------------------------------------------------------------
+	//----------------------------------------------------------------------------
 	//---
 	//--- Service
 	//---
-	//-----------------------------------------------------------------------------
+	//----------------------------------------------------------------------------
 
 	public Element exec(Element params, ServiceContext context) throws Exception
 	{
 		String uploadDir = context.getUploadDir();
 
-		String id     = Util.getParam(params, Params.ID);
-		String ref    = Util.getParam(params, Params.REF);
-		String fname  = Util.getParam(params, Params.FNAME);
-		String access = Util.getParam(params, Params.ACCESS);
+		String id     		= Util.getParam(params, Params.ID);
+		String ref    		= Util.getParam(params, Params.REF);
+		String access 		= Util.getParam(params, Params.ACCESS);
+		String overwrite	= Util.getParam(params, Params.OVERWRITE, "no");
 
 		Lib.resource.checkEditPrivilege(context, id);
+
+		// get info to log the upload
+
+		UserSession session = context.getUserSession();
+		Dbms dbms = (Dbms) context.getResourceManager().open(Geonet.Res.MAIN_DB);
+		String username = session.getUsername();
+		if (username == null) username = "unknown (this shouldn't happen?)";
+
+		StringBuffer query = new StringBuffer();
+		query.append("SELECT m.id, m.source, m.uuid ");
+		query.append("FROM   Metadata m ");
+		query.append("WHERE  m.id = " + id);
+
+		String siteId = "unknown";
+		String mdUuid = "unknown";
+		Element sites = dbms.select(query.toString());
+		if (sites != null) {
+			for (Iterator i = sites.getChildren().iterator(); i.hasNext(); ) {
+				Element site = (Element)i.next();
+				siteId = site.getChildText("source");
+				mdUuid = site.getChildText("uuid");
+				if (siteId == null) siteId = "unknown";
+				if (mdUuid == null) mdUuid = "unknown";
+			}
+		}
 
 		// move uploaded file to destination directory
 		// note: uploadDir and rootDir must be in the same volume
@@ -77,23 +101,55 @@ public class Upload implements Service
 		File dir = new File(Lib.resource.getDir(context, access, id));
 		dir.mkdirs();
 
-		// move uploaded file to destination directory
+		// Jeeves will place the uploaded file name in the f_{ref} element
+		// we do it this way because Jeeves will sanitize the name to remove
+		// characters that may cause problems
+		Element fnameElem = params.getChild("f_"+ref);
+		String fname = fnameElem.getText();
+		String fsize = fnameElem.getAttributeValue("size");
+		if (fsize == null) fsize="0";
+
+		// get ready to move uploaded file to destination directory
 		File oldFile = new File(uploadDir, fname);
 		File newFile = new File(dir,       fname);
 
+		context.info("Source : "+oldFile.getAbsolutePath());
+		context.info("Destin : "+newFile.getAbsolutePath());
+
+		if (!oldFile.exists()) {
+			throw new Exception("File upload unsuccessful "+oldFile.getAbsolutePath()+" does not exist");
+		}
+
+		// check if file already exists and do whatever overwrite wants
+		if (newFile.exists() && overwrite.equals("no")) {
+			throw new Exception("File upload unsuccessful because "+newFile.getName()+" already exists and overwrite was not permitted");
+		}
+
+	
+		// move uploaded file to destination directory - have two goes
 		if (!oldFile.renameTo(newFile))
 		{
-			context.warning("Cannot move uploaded file");
-			context.warning(" (C) Source : "+oldFile.getAbsolutePath());
-			context.warning(" (C) Destin : "+newFile.getAbsolutePath());
-			oldFile.delete();
-			throw new Exception("Unable to move uploaded file to destination directory");
+			newFile.delete();
+			if (!oldFile.renameTo(newFile))
+			{
+				context.warning("Cannot move uploaded file");
+				context.warning(" (C) Source : "+oldFile.getAbsolutePath());
+				context.warning(" (C) Destin : "+newFile.getAbsolutePath());
+				oldFile.delete();
+				throw new Exception("Unable to move uploaded file to destination directory");
+			}
 		}
+
+		// log the upload
+		context.info("UPLOADED:"+fname+","+id+","+mdUuid+","+context.getIpAddress()+","+username);
+
 		// update the metadata
 		Element elem = new Element("_" + ref);
 		params.addContent(elem);
 		elem.setText(fname);
-		return update.exec(params, context);
+		return new Element("response")
+					.addContent(new Element("fname").setText(fname))
+					.addContent(new Element("fsize").setText(fsize));
 	}
 }
